Format

+++ framework/py/flwr/supercore/state/alembic/versions/rev_2026_05_26_add_run_series.py
@@ -60,7 +60,7 @@ def upgrade() -> None:
         "series_runs",
         sa.Column(
             "id",
-            sa.BigInteger(),
+            sa.Integer(),
             autoincrement=True,
             nullable=False,
         ),
